pyproject.toml
[tool.poetry]
name = "comment-detector"
version = "0.1.0"
description = ""
authors = ["Chuan <kirk336neo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
datetime = "^5.2"
apache-airflow = "^2.7.2"
google-cloud = "^0.34.0"
loguru = "^0.7.2"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
pymongo = "^4.5.0"
fake-useragent = "^1.3.0"
google-oauth2l = "^1.0.2"
google-cloud-logging = "^3.8.0"
google-cloud-storage = "^2.12.0"
lxml = "^4.9.3"
numpy = "^1.26.1"
pandas = "^2.1.1"
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"
openai = "^0.28.1"
jieba = "^0.42.1"
dash = "^2.14.0"
plotly = "^5.17.0"
networkx = "^3.1"
matplotlib = "^3.8.0"
wordcloud = "^1.9.2"
uvicorn = "^0.23.2"
fastapi = "^0.103.2"
apscheduler = "^3.10.4"
dash-bootstrap-components = "^1.5.0"
slowapi = "^0.1.8"
gunicorn = "^21.2.0"
configparser = "^6.0.0"
pytest-cov = "^4.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]